Split annotation reticle tests by ownership
